Remove Travis CI integration
It will be restored when there is a way of writing unit tests in
Ionic 2 for an Angular 2 application. It's frustrating, but there is
still no clear way to do that. Angular 2 and Ionic 2 are in beta,
and they are constantly changing things. I've even heard that Ionic 2
is thinking of replacing Webpack in favor of Gulp. I love Gulp, but
there are lots of changes happening and no standard way of writing
tests. Too sad.